Conflicts:
	CMakeLists.txt
	include/chaiscript/dispatchkit/boxed_number.hpp
	include/chaiscript/language/chaiscript_engine.hpp

+++ include/chaiscript/dispatchkit/boxed_number.hpp
@@ -275,7 +275,7 @@ namespace chaiscript
           } else if (inp_ == typeid(std::int8_t)) {
             return oper_rhs<std::int8_t, false>(t_oper, t_lhs, t_rhs);
           } else if (inp_ == typeid(std::int16_t)) {
-            return oper_rhs<std::int32_t, false>(t_oper, t_lhs, t_rhs);
+            return oper_rhs<std::int16_t, false>(t_oper, t_lhs, t_rhs);
           } else if (inp_ == typeid(std::int32_t)) {
             return oper_rhs<std::int32_t, false>(t_oper, t_lhs, t_rhs);
           } else if (inp_ == typeid(std::int64_t)) {
